Fix undefined type RealizedViewType in generated source.
This seems to be a quirk with TS that I'm too new to
solve. Technically the prior solution should have resulted in no
references to the enum in the generated code but alas, there were
some. This ensures the enum is propertly constructed to avoid it. I
couldn't find a way to get rid of them.